chore: rename git_push_stack to mergify_cli
